Mark non-English resources as so
As the majority of the resources are in English, it would make sense to have the ones not so marked clearly.

I spent 5 minutes trying to swap the language in the resource only to find that there wasn't an English version available. If this was marked it could clear up some confusion.